Recognize COMPOSE_TLS_VERSION env var in tls_config_from_options
